Move files around and fix parseFromDocument tests
